auth_api/requirements.txt
alembic==1.8.0
apispec==5.2.2
apispec-webframeworks==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.2
blinker==1.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
flasgger==0.9.5
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.4.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
gevent==21.12.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
mistune==2.0.2
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
packaging==21.3
passlib==1.7.4
protobuf==4.21.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.37
thrift==0.16.0
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.17.6
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
